add function insert bookImage

+++ bookBoard/src/main/java/bookBoard/dto/BookBoardImgDto.java
@@ -1,7 +1,7 @@
 package bookBoard.dto;
 
 public class BookBoardImgDto {
-	private int imageId;
-	private int bookId;
+	private long imageId;
+	private long bookId;
 	private String imageUrl;
 }
